requirements.txt
albucore==0.0.23
albumentations==2.0.5
annotated-types==0.7.0
asgiref==3.8.1
attrs==25.3.0
Brotli==1.1.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.3.1
cryptography==44.0.2
cycler==0.12.1
Cython==3.0.12
defusedxml==0.7.1
dj-database-url==2.3.0
Django==5.1.7
django-cors-headers==4.7.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
djoser==2.3.1
drf-spectacular==0.28.0
drf-yasg==1.21.10
easydict==1.13
flatbuffers==25.2.10
fonttools==4.56.0
gunicorn==23.0.0
h11==0.14.0
humanfriendly==10.0
idna==3.10
imageio==2.37.0
inflection==0.5.1
insightface==0.7.3
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.8
lazy_loader==0.4
lz4==4.4.3
matplotlib==3.10.1
mpmath==1.3.0
mtcnn==1.0.0
networkx==3.4.2
numpy==2.2.4
oauthlib==3.2.2
onnx==1.17.0
onnxruntime==1.21.0
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
packaging==24.2
pillow==11.1.0
prettytable==3.16.0
protobuf==6.30.2
psycopg2-binary==2.9.10
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.9.0
pyparsing==3.2.3
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python3-openid==3.2.0
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.24.0
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.2
simsimd==6.2.1
six==1.17.0
social-auth-app-django==5.4.3
social-auth-core==4.5.6
sqlparse==0.5.3
stringzilla==3.12.3
sympy==1.13.3
threadpoolctl==3.6.0
tifffile==2025.3.13
tqdm==4.67.1
typing_extensions==4.13.0
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.0
wcwidth==0.2.13
whitenoise==6.9.0
